fix scenarios validation
